+++ NEWS.md
@@ -1,8 +1,11 @@
 # mixvlmc (development version)
 ## Breaking changes
-* defaults for `draw.covlmc()` have changed:
+* defaults for `draw.covlmc()` have changed (issue #66):
   * `p_value` is now `FALSE` by default
   * the characters used to separate levels from models is now `" ~ "`
+  * temporal blocks are now separated by default by the `time_sep` member
+    in the `control` list. The default value is set to `" | "` and the
+    `time_sep` parameter has been removed from `draw.covlmc()`
 
 # mixvlmc 0.2.1
 This release has no visible changes and fixes internal issues: